support ObjectMap
